+++ blockchain/src/blockchain/accounts.rs
@@ -37,7 +37,9 @@ impl Blockchain {
 
                 // Check if the receipts contain an error.
                 if let Err(e) = batch_info {
-                    return Err(PushError::AccountsError(e));
+                    // TODO: This is due to a dependency between blockchain and accounts that is not wasm friendly
+                    log::error!("Received accounts error while committing accounts: {}", e);
+                    return Err(PushError::AccountsError);
                 }
 
                 // Macro blocks are final and receipts for the previous batch are no longer necessary
@@ -99,7 +101,9 @@ impl Blockchain {
                     Ok(batch_info) => batch_info,
                     Err(e) => {
                         // Check if the receipts contain an error.
-                        return Err(PushError::AccountsError(e));
+                        // TODO: This is due to a dependency between blockchain and accounts that is not wasm friendly
+                        log::error!("Received accounts error while committing accounts: {}", e);
+                        return Err(PushError::AccountsError);
                     }
                 };
