+++ blockchain/src/blockchain/history_sync.rs
@@ -229,7 +229,12 @@ impl Blockchain {
                 txn.abort();
                 #[cfg(feature = "metrics")]
                 this.metrics.note_invalid_block();
-                return Err(PushError::AccountsError(e));
+                // TODO: This is due to a dependency between blockchain and accounts that is not wasm friendly
+                log::error!(
+                    "Received accounts error while extending history sync: {}",
+                    e,
+                );
+                return Err(PushError::AccountsError);
             }
         }
         this.state.accounts.finalize_batch(&mut txn);
